Project/requirements.txt
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.0

Flask==3.0.0
matplotlib==3.7.4
opencv-python==4.9.0.80
numpy==1.24.3
Pillow==10.1.0
pandas==2.0.3
tk==0.1.0
tensorflow==2.15.0
keras==2.15.0